Move to connections dir
